DOMA-1833 renamed filters to fields in client schema
